Options and Bug Fixes
Added the ability to change the save path for the Daily Worksheets.
This in turn allows the default path to be set to a network location
allowing multiple computers to look at the same worksheet.

Also fixed a bug where the credit-card totals weren't being immediately
calculated into the statistics.